pyproject.toml
[project]
name = "nexa-analyzer-api"
version = "0.1.0"
description = "Nexa - Client's Context Analyzer backend API"
authors = [ { name = "Franco Gaytan" } ]
readme = "README.md"
requires-python = ">=3.13"
license = "Proprietary"

# Runtime dependencies (mirroring requirements.txt)
dependencies = [
    "fastapi==0.115.2",
    "uvicorn[standard]==0.30.1",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "httpx==0.27.2",
    "python-multipart==0.0.9",
    "pypdf==5.0.1",
    "python-docx==1.1.2",
    "beautifulsoup4==4.12.3",
    "requests==2.32.3",
    "openai==1.48.0",
    # "tiktoken==0.7.0"  # removed (needs Rust; re-add when Rust toolchain available)
        "pytest==8.3.3",
]

[tool.uv]
# Optional settings for uv (can be extended later)
cache-dir = ".uv-cache"

[tool.uv.sources]
# If you later host internal packages, define source indexes here.

[tool.uv.workspace]
members = []

[project.optional-dependencies]
dev = [
    "pytest==8.3.3"
]

[tool.pytest.ini_options]
pythonpath = ["."]
